requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow>=12.0
pvlib>=0.10
requests>=2.31
python-dotenv>=1.0
scikit-learn>=1.3
xgboost>=1.7
streamlit>=1.28
plotly>=5.17
matplotlib>=3.7
seaborn>=0.12